Assorted changes
