--- constructorio-client/src/test/java/io/constructor/client/ConstructorIOTest.java
@@ -1,218 +1,0 @@
-package io.constructor.client;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
-import org.junit.BeforeClass;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-
-import okhttp3.OkHttpClient;
-import okhttp3.logging.HttpLoggingInterceptor;
-import okhttp3.logging.HttpLoggingInterceptor.Level;
-
-public class ConstructorIOTest {
-
-    @Rule
-    public ExpectedException thrown = ExpectedException.none();
-
-    @BeforeClass
-    public static void setupHTTPLogging() throws Exception {
-        HttpLoggingInterceptor logger = new HttpLoggingInterceptor();
-        logger.setLevel(Level.NONE);
-        OkHttpClient client = ConstructorIO.getClient();
-        OkHttpClient newClient = client.newBuilder().addInterceptor(logger).build();
-        ConstructorIO.setClient(newClient);
-    }
-
-    @BeforeClass
-    public static void setupItemToTrack() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        ConstructorItem item = new ConstructorItem("Stanley_Steamer");
-        item.setId("Stanley1");
-        item.setUrl("https://constructor.io/products/Stanley1");
-        constructor.addOrUpdateItem(item, "Products");
-        Thread.sleep(2000);
-    }
-
-    @Test
-    public void newShouldSetApiToken() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", true, null);
-        assertEquals("api token should be set", constructor.apiToken, "boinkaToken");
-    }
-
-    @Test
-    public void newShouldSetApiKey() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", true, null);
-        assertEquals("api key should be set", constructor.apiKey, "doinkaKey");
-    }
-
-    @Test
-    public void newShouldSetProtocol() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", false, null);
-        assertEquals("protocol should be set", constructor.protocol, "http");
-    }
-
-    @Test
-    public void newShouldSetHostname() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", false, "com.cnstrc.ac");
-        assertEquals("host should be set", constructor.host, "com.cnstrc.ac");
-    }
-
-    @Test
-    public void newShouldSetHostnameDefault() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", false, null);
-        assertEquals("host should be set to default", constructor.host, "ac.cnstrc.com");
-    }
-
-    @Test	
-    public void makeUrlShouldReturnUrl() throws Exception {	
-        ConstructorIO constructor = new ConstructorIO("boinka", "doinka", true, null);
-        String generatedUrl = constructor.makeUrl("v1/test");
-        assertEquals("make url should make urls", generatedUrl, "https://ac.cnstrc.com/v1/test?key=doinka&c=ciojava-3.3.0");
-    }
-
-    @Test
-    public void verifyShouldReturnTrueWithValidKeyTokenPair() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        assertEquals("verify should return true for testing key/pair", constructor.verify(), true);
-    }
-
-    @Test
-    public void addItemShouldReturnTrue() throws Exception {
-      ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-      ConstructorItem item = Utils.createProductItem();
-      assertTrue("addition succeeds", constructor.addItem(item, "Products"));
-    }
-
-    @Test
-    public void addOrUpdateItemShouldReturnTrue() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        ConstructorItem item = Utils.createProductItem();
-        assertTrue("upsert succeeds", constructor.addOrUpdateItem(item, "Products"));
-    }
-
-    @Test
-    public void addBatchShouldReturnTrue() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        ConstructorItem[] items = {
-            Utils.createProductItem(),
-            Utils.createProductItem(),
-            Utils.createProductItem()
-        };
-        assertTrue("batch addition succeeds", constructor.addItemBatch(items, "Products"));
-    }
-
-    @Test
-    public void addOrUpdateBatchShouldReturnTrue() throws Exception {
-      ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-      ConstructorItem[] items = {
-        Utils.createProductItem(),
-        Utils.createProductItem(),
-        Utils.createProductItem()
-      };
-      assertTrue("batch upsert succeeds", constructor.addOrUpdateItemBatch(items, "Products"));
-    }
-
-    @Test
-    public void modifyShouldReturnTrue() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        ConstructorItem itemOld = Utils.createProductItem();
-        constructor.addItem(itemOld, "Products");
-        Thread.sleep(2000);
-
-        ConstructorItem itemNew = new ConstructorItem(itemOld.getItemName());
-        itemNew.setUrl(itemOld.getUrl());
-        itemNew.setSuggestedScore(1337);
-        assertTrue("modify succeeds", constructor.modifyItem(itemOld, "Products", itemOld.getItemName()));
-    }
-
-    @Test
-    public void removeShouldReturnTrue() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        ConstructorItem item = Utils.createProductItem();
-        constructor.addItem(item, "Products");
-        Thread.sleep(2000);
-
-        assertTrue("remove succeeds", constructor.removeItem(item, "Products"));
-    }
-
-    @Test
-    public void removeBatchShouldReturnTrue() throws Exception {
-      ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-      ConstructorItem[] items = {
-        Utils.createProductItem(),
-        Utils.createProductItem(),
-        Utils.createProductItem()
-      };
-      assertTrue("batch removal succeeds", constructor.removeItemBatch(items, "Products"));
-    }
-
-    @Test
-    public void autocompleteShouldReturnAResult() throws Exception {
-      ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-      UserInfo userInfo = new UserInfo(3, "c62a-2a09-faie");
-      AutocompleteResponse result = constructor.autocomplete("Stanley", userInfo);
-      assertTrue("autocomplete succeeds", result.getResultId() != null);
-    }
-
-    @Test
-    public void autocompleteShouldReturnAResultWithNullUserInfo() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        AutocompleteResponse result = constructor.autocomplete("Stanley", null);
-        assertTrue("autocomplete succeeds", result.getResultId() != null);
-    }
-
-    @Test
-    public void trackConversionShouldReturnTrue() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        UserInfo userInfo = new UserInfo(3, "c62a-2a09-faie");
-        assertTrue("conversion without params succeeds", constructor.trackConversion("Stanley_Steamer", "Products", "Stanley1", "$1.99", userInfo));
-    }
-
-    @Test
-    public void trackConversionShouldReturnTrueWithNullUserInfo() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        assertTrue("conversion without params succeeds", constructor.trackConversion("Stanley_Steamer", "Products", "Stanley1", "$1.99", null));
-    }
-
-    @Test
-    public void trackSearchShouldReturnTrue() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        UserInfo userInfo = new UserInfo(3, "c62a-2a09-faie");
-        assertTrue("search without params succeeds", constructor.trackSearch("Stanley_Steamer", 22, userInfo));
-    }
-
-    @Test
-    public void trackSearchShouldReturnTrueWithNullUserInfo() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        assertTrue("search without params succeeds", constructor.trackSearch("Stanley_Steamer", 22, null));
-    }
-
-    @Test
-    public void trackClickThroughShouldReturnTrue() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        UserInfo userInfo = new UserInfo(3, "c62a-2a09-faie");
-        assertTrue("search without params succeeds", constructor.trackClickThrough("Stanley_Steamer", "Products", "Stanley1", userInfo));
-    }
-
-    @Test
-    public void trackClickThroughShouldReturnTrueWithNullUserInfo() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        assertTrue("search without params succeeds", constructor.trackClickThrough("Stanley_Steamer", "Products", "Stanley1", null));
-    }
-
-    @Test
-    public void getVersionShouldReturnClientVersion() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        assertEquals("grabs version from pom.xml", constructor.getVersion(), "ciojava-3.3.0");
-    }
-
-    @Test
-    public void serializeUserInfoShouldReturnQueryString() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        UserInfo userInfo = new UserInfo(3, "c62a-2a09-faie");
-        assertEquals("serializes user info into query string", constructor.serializeUserInfo(userInfo), "&s=3&i=c62a-2a09-faie");
-    }
-}
